Make health circle heal player on Update()

+++ ProjectMagma/Assets/Scripts/playerController.cs
@@ -482,7 +482,7 @@ public class playerController : MonoBehaviour, IDamage
     //    updatePlayerUI();
     //}
 
-    public void Heal(int value)
+    public void Heal(float value)
     {
         Health += value;
         Healed?.Invoke();
